Generate bindings in build.rs to fix cross-compilation
The pre-generated bindings are not working on the platforms that is not the same platform these bindings are generated at:
[/src/lapack.rs](https://github.com/numrs/lapack-sys/blob/6f42e18ff245e0b53527a24211af1f640b4b637d/src/lapack.rs)

In particular, this is failing during pretty much any attempt at cross-compilation. And might cause problems diring native compilation on different platform as well. For example, when cross-compiling to aarch64-unknown-linux-gnu, the defaut C integer types are different from what's on x86_64:
[build.log.txt](https://github.com/user-attachments/files/17705378/build.log.txt)

The [preferred approach[(https://rust-lang.github.io/rust-bindgen/library-usage.html) nowadays is to generate bindings at build-time and then programmatically paste them into lib.rs. This way they come out with all the specifics required for the target platform.

Here I:
- add `build.rs` script which uses bindgen library to generate `$OUTPUT_DIR/bindings.rs` (in the `target/` directory)
- add the functions starting with underscore to the allowlist (`.allowlist_function("^.*_$")`)
-`include!()` the `$OUTPUT_DIR/bindings.rs`into `lib.rs`
- remove the `generate.sh` script (the `build.rs` is now doing it's job)
- remove pre-generated bindings